Refactor parser functions
